Fix self-signed root linting and move utility functions into cert.go
