+++ cmd/ceremony/main_test.go
@@ -1,43 +1,8 @@
 package main
 
 import (
-	"strings"
 	"testing"
 )
-
-func TestCheckOutputFileSucceeds(t *testing.T) {
-	dir := t.TempDir()
-	err := checkOutputFile(dir+"/example", "foo")
-	if err != nil {
-		t.Fatal(err)
-	}
-}
-
-func TestCheckOutputFileEmpty(t *testing.T) {
-	err := checkOutputFile("", "foo")
-	if err == nil {
-		t.Fatal("expected error, got none")
-	}
-	if err.Error() != "outputs.foo is required" {
-		t.Fatalf("wrong error: %s", err)
-	}
-}
-
-func TestCheckOutputFileExists(t *testing.T) {
-	dir := t.TempDir()
-	filename := dir + "/example"
-	err := writeFile(filename, []byte("hi"))
-	if err != nil {
-		t.Fatal(err)
-	}
-	err = checkOutputFile(filename, "foo")
-	if err == nil {
-		t.Fatal("expected error, got none")
-	}
-	if !strings.Contains(err.Error(), "already exists") {
-		t.Fatalf("wrong error: %s", err)
-	}
-}
 
 func TestKeyGenConfigValidate(t *testing.T) {
 	cases := []struct {
